// src/app/shared/model/VendorGroupObj.Model.ts
export class VendorGroupObj {
    VendorGrpId: any;
    VendorGrpCode: any;
    VendorGrpName: any;
    VenderGrpDesc: any;
    MrVendorCategoryCode: any;
    MrVendorTypeCode: any;
    IsActive: any;
    RowVersion: any;
    VendorId: any;
    
    constructor(){this.VendorGrpId = "0", this.RowVersion = "" ,this.VendorId ="0" }
}